+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v7.9.4-4] - 2021-02-18
+### Changed
+- Members of the CES administrator group receive project admin permissions for new projects (#3)
+- CES_ADMIN group can be enabled to administer all projects using the key `amend_projects_with_ces_admin_permissions`
+(see `dogu.json` for details) (#3)
+
 ## [v7.9.4-3] - 2021-02-01
 ### Fixed
 - pass truststore as jvm options to compute engine (#56)